% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_helpers.R
\name{browseRC}
\alias{browseRC}
\title{Browse a gene on Ribocrypt webpage}
\usage{
browseRC(
  symbol = NULL,
  gene_id = NULL,
  tx_id = NULL,
  exp = "all_merged-Homo_sapiens_modalities",
  libraries = NULL,
  host = "https://ribocrypt.org",
  plot_on_start = TRUE,
  frames_type = "columns",
  kmer = 1,
  browser = getOption("browser")
)
}
\arguments{
\item{symbol}{gene symbol, default NULL}

\item{gene_id}{gene symbol, default NULL}

\item{tx_id}{gene symbol, default NULL}

\item{exp}{experiment name, default "all_merged-Homo_sapiens_modalities"}

\item{libraries}{NULL, default to first in experiment, c("RFP","RNA") would add RNA to default.}

\item{host}{url, default "https://ribocrypt.org". Set to localhost for local version.}

\item{plot_on_start}{logical, default TRUE. Plot gene when opening browser.}

\item{frames_type}{"columns"}

\item{kmer}{default 1 (kmer window to smear out reads)}

\item{browser}{getOption("browser")}
}
\value{
browseURL, opens browse with page
}
\description{
Can also disply local RiboCrypt app
}
