% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_helpers.R
\name{collection_dir_from_exp}
\alias{collection_dir_from_exp}
\title{Get collection directory}
\usage{
collection_dir_from_exp(df, must_exists = FALSE)
}
\arguments{
\item{df}{ORFik experiment}

\item{must_exists}{logical, stop if dir does not exists}
}
\value{
file.path(resFolder(df), "collection_tables")
}
\description{
Get collection directory
}
\examples{
df <- ORFik.template.experiment()
collection_dir_from_exp(df)

}
