% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_helpers.R
\name{collection_path_from_exp}
\alias{collection_path_from_exp}
\title{Get collection directory}
\usage{
collection_path_from_exp(
  df,
  id,
  must_exists = TRUE,
  output_dir = collection_dir_from_exp(df, must_exists)
)
}
\arguments{
\item{df}{ORFik experiment}

\item{id}{character, transcript ids}

\item{must_exists}{logical, stop if dir does not exists}

\item{output_dir}{= collection_dir_from_exp(df, must_exists)}
}
\value{
file.path(resFolder(df), "collection_tables")
}
\description{
Get collection directory
}
