% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_helpers.R
\name{collection_to_wide}
\alias{collection_to_wide}
\title{Cast a collection table to wide format}
\usage{
collection_to_wide(table, value.var = "logscore")
}
\arguments{
\item{table}{a data.table in long format}

\item{value.var}{which column to use as scores, default "logscore"}
}
\value{
a table in wide format
}
\description{
Cast a collection table to wide format
}
