% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_helpers.R
\name{compute_collection_table}
\alias{compute_collection_table}
\title{Get collection table normalized in wide format}
\usage{
compute_collection_table(
  path,
  lib_sizes,
  df,
  metadata_field,
  normalization,
  kmer,
  metadata
)
}
\arguments{
\item{path}{the path to gene counts}

\item{lib_sizes}{named integer vector, default NULL. If given will do a pre tpm normalization
for full library sizes}

\item{normalization}{a character string, which mode, for options see ?RiboCrypt:::normalizations}

\item{kmer}{integer, default 1L (off), if > 1 will smooth out signal with sliding window size kmer.}
}
\value{
a data.table in wide format
}
\description{
Get collection table normalized in wide format
}
