% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_plots.R
\name{createSeqPanelPattern}
\alias{createSeqPanelPattern}
\title{Create sequence panel for RiboCrypt}
\usage{
createSeqPanelPattern(
  sequence,
  start_codons = "ATG",
  stop_codons = c("TAA", "TAG", "TGA"),
  frame = 1,
  custom_motif = NULL
)
}
\arguments{
\item{start_codons}{character vector, default "ATG"}

\item{stop_codons}{character vector, default c("TAA", "TAG", "TGA")}

\item{custom_motif}{character vector, default NULL.}
}
\value{
a ggplot object
}
\description{
Create sequence panel for RiboCrypt
}
\keyword{internal}
