% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_plot}
\alias{DEG_plot}
\title{Differential expression plots (1D or 2D)}
\usage{
DEG_plot(dt, draw_non_regulated = FALSE)
}
\arguments{
\item{dt}{a data.table with results from a differential
expression run. Normally from: \code{ORFik::DTEG.analysis(df1, df2)}}

\item{draw_non_regulated}{logical, default FALSE.
Should non-regulated rows be included in the plot?
Will make the plot faster to render if skipped (FALSE)}
}
\value{
plotly object
}
\description{
Gives you interactive 1D or 2D DE plots
}
\examples{
df <- ORFik.template.experiment()
dt <- DTEG.analysis(df[df$libtype == "RFP",], df[df$libtype == "RNA",],
                    output.dir = NULL)
#DEG_plot(dt, draw_non_regulated = TRUE)
}
