% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots.R
\name{distanceToFollowing}
\alias{distanceToFollowing}
\title{Distance to following range}
\usage{
distanceToFollowing(grl, grl2 = grl, ignore.strand = FALSE)
}
\arguments{
\item{grl}{a GRangesList}

\item{grl2}{a GRangesList, default 'grl'}

\item{ignore.strand}{logical, default FALSE}
}
\value{
numeric vector of distance
}
\description{
Distance to following range
}
