% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{fetch_JS_seq}
\alias{fetch_JS_seq}
\title{Fetch Javascript sequence}
\usage{
fetch_JS_seq(
  target_seq,
  nplots,
  distance = 50,
  display_dist,
  aa_letter_code = "one_letter",
  input_id
)
}
\arguments{
\item{target_seq}{the target sequence}

\item{nplots}{number of plots}

\item{distance}{numeric, default 50.}

\item{display_dist}{display distance}

\item{aa_letter_code}{"one_letter"}

\item{input_id}{shiny id of the object}
}
\value{
a list of 2 lists, the nt list (per frame, total 3)
 and AA list (per frame, total 3)
}
\description{
Fetch Javascript sequence
}
