% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3dbeacons_utils.R
\name{fetch_summary}
\alias{fetch_summary}
\title{Fetch summary of uniprot id}
\usage{
fetch_summary(qualifier, provider = "alphafold")
}
\arguments{
\item{qualifier}{uniprot ids}

\item{provider}{"pdbe", alternatives: "alphafold", "all"}
}
\value{
a character of json
}
\description{
Fetch summary of uniprot id
}
