% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_plots.R
\name{geneTrackLayer}
\alias{geneTrackLayer}
\title{How many rows does the gene track need}
\usage{
geneTrackLayer(grl)
}
\arguments{
\item{grl}{a GRangesList}
}
\value{
numeric, the track row index
}
\description{
How many rows does the gene track need
}
\keyword{internal}
