% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megabrowser_internal.R
\name{get_megabrowser_annotation_plot}
\alias{get_megabrowser_annotation_plot}
\title{Full plot for allsamples browser}
\usage{
get_megabrowser_annotation_plot(
  id,
  df,
  summary = TRUE,
  annotation = TRUE,
  region_type,
  plotType = "plotly",
  tx_annotation,
  display_region,
  cds_annotation,
  custom_regions = NULL,
  viewMode,
  collapse_intron_flank,
  templates = NULL
)
}
\arguments{
\item{id}{id of transcript}

\item{df}{ORFik experiment}

\item{summary}{logical, default TRUE (add top plot)}

\item{annotation}{logical, default TRUE (add bottom annotation track)}

\item{region_type}{character, "what is the coverage region?" Usually full mrna:
"mrna" or "leader+cds".}

\item{plotType}{= "plotly",}

\item{tx_annotation}{a GRangesList of tx annotation}

\item{display_region}{a GRangesList of display region}

\item{cds_annotation}{a GRangesList of cds annotation}

\item{custom_regions}{optional custom annotation regions to draw.}

\item{viewMode}{character, "tx" or "genomic"}

\item{collapse_intron_flank}{integer, if TRUE and viewMode genomic, collapse
introns to this max size.}

\item{templates}{optional plotly template bundle.}
}
\value{
a cowplot grub
}
\description{
Full plot for allsamples browser
}
