% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_getters.R
\name{getCoverageProfile}
\alias{getCoverageProfile}
\title{Get coverage profile}
\usage{
getCoverageProfile(grl, reads, kmers = 1, kmers_type = "mean")
}
\arguments{
\item{grl}{a GRangesList}

\item{reads}{GRanges}

\item{kmers}{1}

\item{kmers_type}{"mean"}
}
\value{
data.table of coverage
}
\description{
Get coverage profile
}
\keyword{internal}
