% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{getIndexes}
\alias{getIndexes}
\title{Get index}
\usage{
getIndexes(ref_granges)
}
\arguments{
\item{ref_granges}{a GRanges object}
}
\value{
integer vector, indices
}
\description{
Get index
}
\keyword{internal}
