% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_helpers.R
\name{ggplotlyHover}
\alias{ggplotlyHover}
\title{Call ggplotly with hoveron defined}
\usage{
ggplotlyHover(x, ...)
}
\arguments{
\item{x}{a a ggplot argument}

\item{...}{additional arguments for ggplotly}
}
\value{
a ggplotly object
}
\description{
Call ggplotly with hoveron defined
}
