% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_helpers.R
\name{load_collection}
\alias{load_collection}
\title{Load a ORFik collection table}
\usage{
load_collection(path)
}
\arguments{
\item{path}{the path to gene counts}
}
\value{
a data.table in long format
}
\description{
Load a ORFik collection table
}
