% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_helpers.R
\name{make_rc_url}
\alias{make_rc_url}
\title{Create URL to browse a gene on Ribocrypt webpage}
\usage{
make_rc_url(
  symbol = NULL,
  gene_id = NULL,
  tx_id = NULL,
  exp = "all_merged-Homo_sapiens_modalities",
  libraries = NULL,
  extendLeaders = 0,
  extendTrailers = 0,
  viewMode = FALSE,
  other_tx = FALSE,
  host = "https://ribocrypt.org",
  plot_on_start = TRUE,
  frames_type = "columns",
  kmer = 1
)
}
\arguments{
\item{symbol}{gene symbol, default NULL}

\item{gene_id}{gene symbol, default NULL}

\item{tx_id}{gene symbol, default NULL}

\item{exp}{experiment name, default "all_merged-Homo_sapiens_modalities"}

\item{libraries}{NULL, default to first in experiment, c("RFP","RNA") would add RNA to default.}

\item{viewMode}{FALSE (transcript view), TRUE gives genomic.}

\item{other_tx}{FALSE, show all other annotation in region (isoforms etc.)}

\item{host}{url, default "https://ribocrypt.org". Set to localhost for local version.}

\item{plot_on_start}{logical, default TRUE. Plot gene when opening browser.}

\item{frames_type}{"columns"}
}
\value{
browseURL, opens browse with page
}
\description{
Can also make url for local RiboCrypt app'
On the actuall app, the function make_url_from_inputs is used on
the shiny reactive input object. This one is for manual use.
}
