% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{matchMultiplePatterns}
\alias{matchMultiplePatterns}
\title{Match multiple patterns}
\usage{
matchMultiplePatterns(patterns, Seq)
}
\arguments{
\item{patterns}{character}

\item{Seq}{a DNAStringSet}
}
\value{
integer vector, indices (named with pattern hit)
}
\description{
Match multiple patterns
}
\keyword{internal}
