% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{matchToGRanges}
\alias{matchToGRanges}
\title{Match to GRanges}
\usage{
matchToGRanges(matches, ref_granges)
}
\arguments{
\item{matches}{integer vector, indices}

\item{ref_granges}{GRanges}
}
\value{
GRanges object
}
\description{
Match to GRanges
}
\keyword{internal}
