% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_default_ORFikExp.R
\name{multiOmicsPlot_ORFikExp}
\alias{multiOmicsPlot_ORFikExp}
\title{Multi-omics plot using ORFik experiment input}
\usage{
multiOmicsPlot_ORFikExp(
  target_range,
  annotation = target_range,
  df,
  reference_sequence = findFa(df),
  reads = outputLibs(df, type = "pshifted", output.mode = "envirlist"),
  withFrames = libraryTypes(df, uniqueTypes = FALSE) \%in\% c("RFP", "RPF", "LSU"),
  frames_type = "lines",
  colors = NULL,
  kmers = NULL,
  ylabels = bamVarName(df),
  proportions = NULL,
  width = NULL,
  height = NULL,
  plot_name = "default",
  plot_title = NULL,
  display_sequence = FALSE,
  annotation_names = NULL
)
}
\arguments{
\item{target_range}{the target region, a \code{\link{GRangesList}} or \code{\link{GRanges}} object}

\item{annotation}{the whole annotation which your target is a subset,
a \code{\link{GRangesList}} or \code{\link{GRanges}} object}

\item{df}{an ORFik \code{\link{experiment}} or a list containing ORFik experiments.
Usually a list when you have split Ribo-seq and RNA-seq etc.}

\item{reference_sequence}{the genome reference, default findFa(df)}

\item{reads}{the NGS libraries, as a list of \code{\link{GRanges}} with or without score column for replicates.
Default: \code{outputLibs(df, type = "pshifted", output.mode = "envirlist")}}

\item{withFrames}{a logical vector, default
\code{libraryTypes(df, uniqueTypes = FALSE) \%in\% c("RFP", "RPF", "LSU")}
Alternative: a length 1 or same length as list length of "reads" argument.}

\item{frames_type}{character, default "lines". Alternative:\cr
- columns \cr
- stacks \cr
- area}

\item{colors}{character, default NULL (automatic colouring). If "withFrames" argument
is TRUE, colors are set to to c("red", "green", "blue") for the 3 frames.
Alternative:  Character vector of length 1 or length of "reads" list argument.}

\item{kmers}{numeric (integer), bin positions into kmers.}

\item{ylabels}{character, default \code{bamVarName(df)}. Name of libraries in "reads" list arugment.}

\item{proportions}{numeric, default NULL. Width of plot.}

\item{width}{numeric, default NULL. Width of plot.}

\item{height}{numeric, default NULL. Height of plot.}

\item{plot_name}{= character, default "default" (will create name from target_range name).
Alternative: custom name for region.}

\item{plot_title}{character, default NULL. A title for plot.}

\item{display_sequence}{logical, default FALSE. If TRUE, display nucleotide sequence in plot.}

\item{annotation_names}{character, default NULL. Alternative naming for annotation.}
}
\value{
the plot object
}
\description{
Customizable html plots for visualizing genomic data.
}
