% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_panels_init.R
\name{organism_input_select}
\alias{organism_input_select}
\title{Select box for organism}
\usage{
organism_input_select(genomes, ns)
}
\arguments{
\item{genomes}{name of genomes, returned from list.experiments()}

\item{ns}{the ID, for shiny session}
}
\value{
selectizeInput object
}
\description{
Select box for organism
}
