% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{RiboCrypt.template.experiment}
\alias{RiboCrypt.template.experiment}
\title{An ORFik experiment to see how it looks}
\usage{
RiboCrypt.template.experiment(as.temp = FALSE)
}
\arguments{
\item{as.temp}{logical, default FALSE, load as ORFik experiment.
If TRUE, loads as data.frame template of the experiment.}
}
\value{
an ORFik experiment
}
\description{
Toy-data created to resemble human genes:\cr
Number of genes: 6\cr
Ribo-seq: 2 libraries
RNA-seq: 2 libraries
CAGE: 1 library
PAS (poly-A): 1 library
}
\examples{
ORFik.template.experiment()
}
\concept{ORFik_experiment}
