% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_registry.R
\name{ribocrypt_access_workspace}
\alias{ribocrypt_access_workspace}
\title{Register a Workspace}
\usage{
ribocrypt_access_workspace(con, id, label = id)
}
\arguments{
\item{con}{Registry connection.}

\item{id}{Immutable workspace ID.}

\item{label}{Display name.}
}
\description{
Register a Workspace
}
