% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboCrypt_app_brochure.R
\name{RiboCrypt_app_brochure}
\alias{RiboCrypt_app_brochure}
\title{Create RiboCrypt app}
\usage{
RiboCrypt_app_brochure(validate.experiments = TRUE)
}
\arguments{
\item{validate.experiments}{logical, default TRUE, set to FALSE
to allow starting the app with malformed experiments, be careful
will crash if you try to load that experiment!}
}
\value{
RiboCrypt shiny app
}
\description{
Create RiboCrypt app
}
