% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_helpers.R
\name{compute_collection_table}
\alias{compute_collection_table}
\title{Get collection table normalized in wide format}
\usage{
compute_collection_table(
  path,
  lib_sizes,
  df,
  metadata_field,
  normalization,
  kmer,
  metadata,
  min_count = 0,
  format = "wide",
  value.var = "logscore",
  as_list = FALSE
)
}
\arguments{
\item{path}{the path to gene counts}

\item{lib_sizes}{named integer vector, default NULL. If given will do a pre tpm normalization
for full library sizes}

\item{df}{the ORFik experiment to load the precomputed collection from.
It must also have defined runIDs() for all samples.}

\item{metadata_field}{the column name in metadata, to select to group on.}

\item{normalization}{a character string, which mode, for options see RiboCrypt:::normalizations}

\item{kmer}{integer, default 1L (off), if > 1 will smooth out signal with sliding window size kmer.}

\item{metadata}{a data.table of metadata, must contain the Run column to
select libraries.}

\item{min_count}{integer, default 0. Minimum counts of coverage over transcript
to be included.}

\item{format}{character, default "wide", alternative "long". The format of
the table output.}

\item{value.var}{which column to use as scores, default "logscore"}

\item{as_list}{logical, default FALSE. Return as list of size 2,
count data.table and metadata data.table Set to TRUE if you need metadata
subset (needed if you subset the table, to get correct matching)}
}
\value{
a data.table in long or wide (default) format, if as list, it is a
list of size 2 (see argument as_list)
}
\description{
Get collection table normalized in wide format
}
