% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_DEG.R
\name{DEG_server}
\alias{DEG_server}
\title{DEG analysis}
\usage{
DEG_server(
  id,
  all_experiments,
  env,
  df,
  experiments,
  libs,
  org,
  gene_name_list,
  rv
)
}
\description{
A contrast compares 2 levels of a factor
Given factor from 9 elements (samples):
  A,A,A,B,B,B,C,C,C
This is a contrast
A + B
A + B + A/B
}
