% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{fetch_JS_seq}
\alias{fetch_JS_seq}
\title{Fetch Javascript sequence}
\usage{
fetch_JS_seq(
  target_seq,
  nplots,
  distance = 250,
  aa_letter_code = "one_letter",
  input_id,
  frame_colors = "R"
)
}
\arguments{
\item{target_seq}{the target sequence}

\item{nplots}{number of plots}

\item{distance}{numeric, default 250 When to show sequence, when x-range is
<= to this number.}

\item{aa_letter_code}{character, default: "one_letter", alternative: three_letters}

\item{input_id}{shiny id of the object}
}
\value{
a list of 2 lists, the nt list (per frame, total 3)
 and AA list (per frame, total 3)
}
\description{
Fetch Javascript sequence
}
