% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_helpers.R
\name{load_collection}
\alias{load_collection}
\title{Load a ORFik collection table}
\usage{
load_collection(path, grl = attr(path, "range"))
}
\arguments{
\item{path}{the path to gene counts}

\item{grl}{a GRangesList, default attr(path, "range"),
for new fst format, which range to get.}
}
\value{
a data.table in long format
}
\description{
Load a ORFik collection table
}
