% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_default.R
\name{multiOmicsPlot_animate}
\alias{multiOmicsPlot_animate}
\title{Multi-omics animation using list input}
\usage{
multiOmicsPlot_animate(
  target_range,
  annotation = target_range,
  reference_sequence,
  reads,
  withFrames = NULL,
  colors = NULL,
  kmers = NULL,
  kmers_type = c("mean", "sum")[1],
  ylabels = NULL,
  proportions = NULL,
  width = NULL,
  height = NULL,
  plot_name = "default",
  plot_title = NULL,
  display_sequence = FALSE,
  annotation_names = NULL,
  start_codons = "ATG",
  stop_codons = c("TAA", "TAG", "TGA"),
  custom_motif = NULL
)
}
\arguments{
\item{target_range}{the target region, a \code{\link[GenomicRanges]{GRangesList}} or \code{\link[GenomicRanges]{GRanges}} object}

\item{annotation}{the whole annotation which your target is a subset,
a \code{\link[GenomicRanges]{GRangesList}} or \code{\link[GenomicRanges]{GRanges}} object}

\item{reference_sequence}{the genome reference,
a \code{\link[Rsamtools]{FaFile}} or \code{\link[Rsamtools]{FaFile}} convertible object}

\item{reads}{the NGS libraries, as a list of \code{\link[GenomicRanges]{GRanges}} with or without score column for replicates.}

\item{withFrames}{a logical vector, default NULL. Alternative: a length 1 or same length as list length of "reads" argument.}

\item{colors}{character, default NULL (automatic colouring). If "withFrames" argument
is TRUE, colors are set to to c("red", "green", "blue") for the 3 frames.
Alternative:  Character vector of length 1 or length of "reads" list argument.}

\item{kmers}{numeric (integer), bin positions into kmers.}

\item{kmers_type}{character, function used for kmers sliding window. default: "mean", alternative: "sum"}

\item{ylabels}{character, default NULL. Name of libraries in "reads" list arugment.}

\item{proportions}{numeric, default NULL. Width of plot.}

\item{width}{numeric, default NULL. Width of plot.}

\item{height}{numeric, default NULL. Height of plot.}

\item{plot_name}{= character, default "default" (will create name from target_range name).
Alternative: custom name for region.}

\item{plot_title}{character, default NULL. A title for plot.}

\item{display_sequence}{logical, default FALSE. If TRUE, display nucleotide sequence in plot.}

\item{annotation_names}{character, default NULL. Alternative naming for annotation.}

\item{start_codons}{character vector, default "ATG"}

\item{stop_codons}{character vector, default c("TAA", "TAG", "TGA")}

\item{custom_motif}{character vector, default NULL.}
}
\value{
the plot object
}
\description{
Multi-omics animation using list input
}
\examples{
library(ORFik)
df <- ORFik.template.experiment()[3,] #Use third library in experiment only
if (requireNamespace("BSgenome.Hsapiens.UCSC.hg19")) {
  cds <- loadRegion(df, "cds")
  multiOmicsPlot_ORFikExp(extendLeaders(extendTrailers(cds[1], 30), 30), df = df,
                        reference_sequence = BSgenome.Hsapiens.UCSC.hg19::Hsapiens,
                        frames_type = "columns")
}
}
