% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboCrypt_app.R
\name{RiboCrypt_app}
\alias{RiboCrypt_app}
\title{Create RiboCrypt app}
\usage{
RiboCrypt_app(validate.experiments = TRUE)
}
\arguments{
\item{validate.experiments}{logical, default TRUE. Check integrity of all experiments
before you start, set to FALSE if you do not care of this to be TRUE.}
}
\value{
RiboCrypt shiny app
}
\description{
Create RiboCrypt app
}
