% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboCrypt_app.R
\name{RiboCrypt_app}
\alias{RiboCrypt_app}
\title{Create RiboCrypt app}
\usage{
RiboCrypt_app(
  validate.experiments = TRUE,
  options = list(launch.browser = ifelse(interactive(), TRUE, FALSE)),
  all_exp = list.experiments(validate = validate.experiments),
  browser_options = c(),
  init_tab_focus = "browser"
)
}
\arguments{
\item{validate.experiments}{logical, default TRUE, set to FALSE
to allow starting the app with malformed experiments, be careful
will crash if you try to load that experiment!}

\item{options}{list of arguments, default
\code{list("launch.browser" = ifelse(interactive(), TRUE, FALSE))}}

\item{all_exp}{a data.table, default:
\code{list.experiments(validate = validate.experiments)}. Which experiments
do you want to allow your app to see, default is all in your system config
path.}

\item{browser_options}{named character vector of browser specific arguments:\cr
- default_experiment : Which experiment to select, default: first one\cr
- default_gene : Which genes to select, default: first one\cr
- default_libs : Which libraries to select: first one, else a single string,
where libs are seperated by "|", like "RFP_WT_r1|RFP_WT_r2".\cr
- default_kmer : K-mer windowing size, default: 1\cr
- default_frame_type : Ribo-seq line type, default: "lines"\cr
- plot_on_start : Plot when starting, default: "FALSE"\cr}

\item{init_tab_focus}{character, default "browser". Which tab to open on
init.}
}
\value{
RiboCrypt shiny app
}
\description{
Create RiboCrypt app
}
\examples{
## Default run
# RiboCrypt_app()
## Plot on start
# RiboCrypt_app(browser_options = c(plot_on_start = "TRUE"))
## Init with an experiment and gene (you must of course have the experiment)

RiboCrypt_app(validate.experiments = FALSE,
       browser_options = c(plot_on_start = "TRUE",
                           default_experiment = "human_all_merged_l50",
                           default_gene = "ATF4-ENSG00000128272"))
}
