% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboCrypt_app.R
\name{RiboCrypt_app}
\alias{RiboCrypt_app}
\title{Create RiboCrypt app}
\usage{
RiboCrypt_app(
  validate.experiments = TRUE,
  options = list(launch.browser = ifelse(interactive(), TRUE, FALSE)),
  all_exp = list.experiments(validate = validate.experiments),
  browser_options = c()
)
}
\arguments{
\item{validate.experiments}{logical, default TRUE, set to FALSE
to allow starting the app with malformed experiments, be careful
will crash if you try to load that experiment!}

\item{options}{list of arguments, default
\code{list("launch.browser" = ifelse(interactive(), TRUE, FALSE))}}

\item{all_exp}{a data.table, default:
\code{list.experiments(validate = validate.experiments)}. Which experiments
do you want to allow your app to see, default is all in your system config
path.}
}
\value{
RiboCrypt shiny app
}
\description{
Create RiboCrypt app
}
\examples{

## To run in RSTUDIO server using ssh
## A proxy url path is made, so we need to assign that
## First run the app, and lcopy url part after port
## should look something like this: /p/3b3a7b68/

}
