% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboCrypt_app_brochure.R
\name{RiboCrypt_app_brochure}
\alias{RiboCrypt_app_brochure}
\title{Create RiboCrypt app}
\usage{
RiboCrypt_app_brochure(
  validate.experiments = TRUE,
  options = list(launch.browser = ifelse(interactive(), TRUE, FALSE)),
  base.url = ""
)
}
\arguments{
\item{validate.experiments}{logical, default TRUE, set to FALSE
to allow starting the app with malformed experiments, be careful
will crash if you try to load that experiment!}

\item{options}{list of arguments, default
\code{list("launch.browser" = ifelse(interactive(), TRUE, FALSE))}}

\item{base.url}{character, default "". To set base url for interactive
session over a connection or localhost, do not touch if you do not know what
you are doing!}
}
\value{
RiboCrypt shiny app
}
\description{
Create RiboCrypt app
}
\examples{

## To run in RSTUDIO server using ssh
## A proxy url path is made, so we need to assign that
## First run the app, and lcopy url part after port
## should look something like this: /p/3b3a7b68/

}
