% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{trimOverlaps}
\alias{trimOverlaps}
\title{Trim overlaps}
\usage{
trimOverlaps(overlaps, target_range)
}
\arguments{
\item{overlaps}{GRanges}

\item{target_range}{GRanges}
}
\value{
GRanges
}
\description{
Trim overlaps
}
\keyword{internal}
